+++ Footballer/Ball.h
@@ -8,12 +8,12 @@ class Ball :
 public:
 	Ball(string aFile);
 	~Ball();
-	void reset(uint aPosX, uint aPosY, uint aAngle);
+	void reset(float aPosX, float aPosY, float aAngle);
 
 private:
-	uint m_angle{ 0 };
-	uint m_posY{ 0 };
-	uint m_posX{ 0 };
+	float m_angle{ 0 };
+	float m_posY{ 0 };
+	float m_posX{ 0 };
 
 	Texture* m_pTexture{ nullptr };
 };
